slightly improved the accuracy of professor searches (blame RMP) and removed event hooks for two commands
It seems like the search engine at RMP is inconsistent with its results, despite having specific inputs. To improve this (sort of), I checked to see if the professor teaches at UW-Madison, rather than checking the fallback boolean (as that can also apply to the UW-Madison profs). Though, this doesn't solve the issue 100% as the search engine itself is a bit wonky.